refactor: split idea blueprint modules
